requirements.txt
gnukek==1.0.0
fastapi==0.85.0
python-multipart==0.0.5
uvicorn==0.18.3
